refactored rendering pipeline by introducing Scene, GameObject, Renderable, Mesh, Vertex, Material and ShaderDefinition types
The Renderer is now generic and can render any kind of Renderable. Every GameObject has a Renderable, which defines how it should be rendered. The Renderable has a Mesh (with vertices) and a Material (with a shader).
A Scene is a kind of Layer which holds GameObjects. The renderer will then select the scenes from the layer stack and render each scene.
The sample wgsl shader is also more generic and draws vertex data from a buffer.